Added cmd+event handler
