+++ .env.example
@@ -1,1 +1,1 @@
-DISCORD_TOKEN=
+TOKEN=
